# portfolio/public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "Minimal"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
...
  "paths": {
...
      "/block": {
          "post": {
              "tags": [
                  "Block"
              ],
              "summary": "Block a user",
              "description": "Block another user. The authenticated user blocks the specified user.",
              "operationId": "blockUser",
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                      "name": "Authorization",
                      "in": "header",
                      "description": "JWT Bearer token",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "name": "body",
                      "in": "body",
                      "description": "User ID to block",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "blocked_id": {
                                  "type": "integer",
                                  "example": 3
                              }
                          }
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successfully blocked the user",
                      "schema": {
                          "type": "object",
                          "properties": {
                              "message": {
                                  "type": "string",
                                  "example": "Người dùng đã bị chặn thành công!"
                              }
                          }
                      }
                  },
                  "400": {
                      "description": "Bad request - User already blocked or trying to block self"
                  },
                  "401": {
                      "description": "Unauthorized - Invalid or missing token"
                  },
                  "500": {
                      "description": "Internal server error"
                  }
              },
              "security": [
                  {
                      "Bearer": []
                  }
              ]
          }
      },
      "/unblock": {
          "post": {
              "tags": [
                  "Block"
              ],
              "summary": "Unblock a user",
              "description": "Unblock a previously blocked user.",
              "operationId": "unblockUser",
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                      "name": "Authorization",
                      "in": "header",
                      "description": "JWT Bearer token",
                      "required": true,
                      "type": "string"
                  },
                  {
                      "name": "body",
                      "in": "body",
                      "description": "User ID to unblock",
                      "required": true,
                      "schema": {
                          "type": "object",
                          "properties": {
                              "blocked_id": {
                                  "type": "integer",
                                  "example": 3
                              }
                          }
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successfully unblocked the user",
                      "schema": {
                          "type": "object",
                          "properties": {
                              "message": {
                                  "type": "string",
                                  "example": "Người dùng đã được bỏ chặn thành công!"
                              }
                          }
                      }
                  },
                  "400": {
                      "description": "Bad request - User not blocked"
                  },
                  "401": {
                      "description": "Unauthorized - Invalid or missing token"
                  },
                  "500": {
                      "description": "Internal server error"
                  }
              },
              "security": [
                  {
                      "Bearer": []
                  }
              ]
          }
      },
      "/blocked-users": {
          "get": {
              "tags": [
                  "Block"
              ],
              "summary": "Get blocked users",
              "description": "Retrieve a list of users blocked by the authenticated user.",
              "operationId": "getBlockedUsers",
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                      "name": "Authorization",
                      "in": "header",
                      "description": "JWT Bearer token",
                      "required": true,
                      "type": "string"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "List of blocked users",
                      "schema": {
                          "type": "array",
                          "items": {
                              "type": "object",
                              "properties": {
                                  "id": {
                                      "type": "integer",
                                      "example": 3
                                  },
                                  "username": {
                                      "type": "string",
                                      "example": "john_doe"
                                  },
                                  "profile_picture": {
                                      "type": "string",
                                      "example": "https://example.com/profile.jpg"
                                  }
                              }
                          }
                      }
                  },
                  "401": {
                      "description": "Unauthorized - Invalid or missing token"
                  },
                  "500": {
                      "description": "Internal server error"
                  }
              },
              "security": [
                  {
                      "Bearer": []
                  }
              ]
          }
      },
...
      "/categories/photos": {
          "get": {
              "tags": [
                  "Category"
              ],
              "summary": "Get Photos by Category Slugs",
              "operationId": "getPhotosByCategorySlugs",
              "produces": [
                  "application/json"
              ],
              "parameters": [
                  {
                      "name": "slugs",
                      "in": "query",
                      "description": "Comma-separated list of category slugs",
                      "required": true,
                      "type": "string"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Photos retrieved successfully",
                      "schema": {
                          "type": "array",
                          "items": {
                              "type": "object",
                              "properties": {
                                  "id": {
                                      "type": "integer"
                                  },
                                  "title": {
                                      "type": "string"
                                  },
                                  "description": {
                                      "type": "string"
                                  },
                                  "upload_date": {
                                      "type": "string",
                                      "format": "date-time"
                                  },
                                  "location": {
                                      "type": "string"
                                  },
                                  "image_url": {
                                      "type": "string"
                                  },
                                  "photo_status": {
                                      "type": "string"
                                  },
                                  "user_id": {
                                      "type": "integer"
                                  },
                                  "category_id": {
                                      "type": "integer"
                                  },
                                  "photo_token": {
                                      "type": "string"
                                  },
                                  "privacy_status": {
                                      "type": "string"
                                  },
                                  "tags": {
                                      "type": "array",
                                      "items": {
                                          "type": "object",
                                          "properties": {
                                              "id": {
                                                  "type": "integer"
                                              },
                                              "name": {
                                                  "type": "string"
                                              }
                                          }
                                      }
                                  },
                                  "user": {
                                      "type": "object",
                                      "properties": {
                                          "id": {
                                              "type": "integer"
                                          },
                                          "username": {
                                              "type": "string"
                                          },
                                          "profile_picture": {
                                              "type": "string"
                                          },
                                          "bio": {
                                              "type": "string"
                                          }
                                      }
                                  }
                              }
                          }
                      }
                  },
                  "404": {
                      "description": "Categories not found"
                  },
                  "500": {
                      "description": "Internal server error"
                  }
              }
          }
      },
...
},
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ]
}